req: use ldd to check whether a soname is provided by a ROS package
The previous approach (using 'rpm -q --whatprovides') did not work
because it required the package that provides the soname to be
installed, which is not the case if it is provided by the package
currently being built. Also, relying on high-level tools such as rpm is
not very reliable, as we only checked whether there is a ROS package
that provides that soname instead of checking which library is actually
used by the dynamic linker.

Instead, check the path that would be loaded for the soname, and add the
soname as a ROS dependency if it is provided by a library in the ROS
path.